requirements.txt
streamlit>=1.30.0
ollama>=0.4.0
feedparser>=6.0.0
python-dateutil>=2.8.0
markdown>=3.5.0
